refactor: adjust formalization for components
