+++ graphiql-app/src/components/index.ts
@@ -1,0 +1,4 @@
+import AuthHeader from './AuthHeader/AuthHeader';
+import NonAuthHeader from './NonAuthHeader/NonAuthHeader';
+
+export { AuthHeader, NonAuthHeader };
