+++ graphiql-app/src/layouts/AuthLayout/AuthLayout.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
 import { Container } from '@mui/material';
+import { AuthHeader } from '../../components';
 
 import styles from './AuthLayout.module.scss';
-import AuthHeader from '../../components/AuthHeader/AuthHeader';
 
 function AuthLayout() {
   return (
