+++ graphiql-app/src/layouts/NonAuthLayout/NonAuthLayout.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
 import { Container } from '@mui/material';
+import { NonAuthHeader } from '../../components';
 
 import styles from './NonAuthLayout.module.scss';
-import NonAuthHeader from '../../components/NonAuthHeader/NonAuthHeader';
 
 function NonAuthLayout() {
   return (
