+++ graphiql-app/src/pages/SignIn/SignIn.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Form from './components/Form';
+import { Form } from './components';
 
 function SignIn() {
   return (
