+++ graphiql-app/src/pages/SignIn/components/index.ts
@@ -1,0 +1,3 @@
+import Form from './Form/Form';
+
+export { Form };
